telaviv_back-main/requirements.txt
APScheduler==3.10.1
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
et_xmlfile==2.0.0
Flask==3.0.3
Flask-Cors==4.0.2
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysql-connector-python==8.2.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
protobuf==4.21.12
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
Werkzeug==3.0.3